fix: 아바타 수정 동작 확인

See merge request s10-final/S10P31E106!115

+++ backend/connectus/spot-service/src/main/java/social/connectus/location/domain/services/SpotService.java
@@ -36,7 +36,7 @@ public class SpotService implements SpotUseCase {
 
     @Override
     public FindNearbyElementResponse findNearbyElement(FindNearbyElementCommand command) {
-        final long RADIUS = 2;
+        final double RADIUS = 2;
         R<QueryResults> resp = milvusPort.select(command.getLongitude(), command.getLatitude());
 //        System.out.println(resp);
